+++ src/main/resources/db/migration/V3__Create_products_table.sql
@@ -1,9 +1,10 @@
 -- V3__Create_products_table
 
 CREATE TABLE products (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(100) NOT NULL,
     price FLOAT NOT NULL,
-    category_id INT NOT NULL,
+    category_id BIGINT NOT NULL,
     CONSTRAINT fk_category FOREIGN KEY (category_id) REFERENCES categories(id)
 );
+
